Basic Feature detection
This commit updates issue #11 and #13.
It gathers the data from the video feed;
Transform it into an OpenCV compatible format;
Apply a mask on the color wanted;
Detected the top and bottom coordinate of the color and assign it into points;
Draw circles to locate the points on the image and display it.